+++ src/MitochondrialDynamics.jl
@@ -4,7 +4,9 @@ using ModelingToolkit
 
 export make_model
 
+include("utils.jl")
 include("rates.jl")
 include("mtk.jl")
+include("model.jl")
 
 end # module
